Added a testing framework and tests for memset and strlen.
All tests can be performed at runtime and the serial port output used to
read the results if required. Obviously, this assumes that the serial port
output code and *printf are all functional.

At the moment it's pretty convoluted and (ab)uses the preprocessor a lot.

Add -D_TESTING to CFLAGS in the Makefile to create a test build.